+++ index.html
@@ -8,7 +8,8 @@
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <link data-trunk rel="icon" href="public/favicon.ico" />
     <link data-trunk rel="scss" href="styles/main.scss" />
-    <link data-trunk rel="tailwind-css" href="styles/output.css" />
+    <link data-trunk rel="tailwind-css" href="styles/tailwind.css"/>
+    <link data-trunk rel="css" href="styles/flags.css" />
     <link data-trunk rel="copy-dir" href="public/images"/>
     <link data-trunk rel="copy-dir" href="public/js/" />
   </head>
