+++ src/components/flag_icon.rs
@@ -2,13 +2,15 @@ use leptos::*;
 
 #[component]
 pub fn FlagIcon(
-    mut country_code: String,
+    country_code: String,
 ) -> impl IntoView {
-    if country_code == "en" {
-        country_code = "gb".to_string();
+    let mut flag_icon = country_code.clone();
+    flag_icon = flag_icon.to_lowercase();
+    if flag_icon == "en" {
+        flag_icon = "gb".to_string();
     }
 
     view! {
-        <span class=format!("fi fis fiCircle inline-block mr-2 flag-icon-{}", country_code) />
+        <span class=format!("fi fis fiCircle inline-block mr-2 fi-{}", flag_icon) />
     }
 }
