+++ src/pages/mod.rs
@@ -6,3 +6,4 @@ mod information;
 mod loading;
 mod settings;
 mod agreement;
+mod updating;
