--- styles/input.css
@@ -1,81 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-
-html, body {
-    background-color: #E5E7EB;
-    height: 100%;
-    width: 100%;
-}
-
-.alert-warning {
-    color: #856404;
-    background-color: #fff3cd;
-    border-color: #ffeeba;
-}
-
-.alert-info {
-    color: #0c5460;
-    background-color: #d1ecf1;
-    border-color: #bee5eb;
-}
-
-.alert-success {
-    color: #155724;
-    background-color: #d4edda;
-    border-color: #c3e6cb;
-}
-
-.alert-danger {
-    color: #721c24;
-    background-color: #f8d7da;
-    border-color: #f5c6cb;
-}
-
-.alert {
-    position: relative;
-    padding: .75rem 1.25rem;
-    margin-bottom: 1rem;
-    border: 1px solid transparent;
-    border-radius: .25rem;
-}
-
-.close {
-    float: right;
-    font-size: 1.5rem;
-    font-weight: 700;
-    line-height: 1;
-    color: #000;
-    text-shadow: 0 1px 0 #fff;
-    opacity: .5;
-    cursor: pointer;
-}
-
-.dropdown-menu {
-    min-width: 0 !important;
-}
-
-.fiCircle {
-    width: 24px !important;
-    height: 24px !important;
-    font-size: 24px !important;
-    border-radius: 100%;
-    border: none;
-    box-shadow: inset 0 0 0 2px rgba(0, 0, 0, .06);
-    background: white;
-}
-
-/* Provided by: https://gedkirkham.dev/blog/post/loading-icon-button-tutorial.html */
-@keyframes rotate {
-    from {transform: rotate(0deg)}
-    to {transform: rotate(360deg)}
-}
-
-.refreshStart {
-    animation-name: rotate;
-    animation-duration: 1s;
-    animation-iteration-count: infinite;
-    animation-timing-function: linear;
-    animation-play-state: running;
-}
