+++ styles/tailwind.css
@@ -1,0 +1,41 @@
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+
+* {
+    font-family: Roboto, sans-serif;
+}
+
+html, body {
+    background-color: #E5E7EB;
+}
+
+html {
+    font-size: 16px;
+    height: 100%;
+    width: 100%;
+}
+
+.fiCircle {
+    width: 24px !important;
+    height: 24px !important;
+    font-size: 24px !important;
+    border-radius: 100%;
+    border: none;
+    box-shadow: inset 0 0 0 2px rgba(0, 0, 0, .06);
+    background: white;
+}
+
+/* Provided by: https://gedkirkham.dev/blog/post/loading-icon-button-tutorial.html */
+@keyframes rotate {
+    from {transform: rotate(0deg)}
+    to {transform: rotate(360deg)}
+}
+
+.refreshStart {
+    animation-name: rotate;
+    animation-duration: 1s;
+    animation-iteration-count: infinite;
+    animation-timing-function: linear;
+    animation-play-state: running;
+}
